refactor: group conversation helpers
