+++ OcchioOnniveggente/src/conversation/state.py
@@ -11,12 +11,12 @@ from openai import OpenAI
 
 # Support import both as ``OcchioOnniveggente.src`` and ``src`` package
 try:  # pragma: no cover - optional relative import
-    from ..DataBase.conversation_store import ConversationStore
+    from ...DataBase.conversation_store import ConversationStore
 except ImportError:  # pragma: no cover - when ``src`` is top-level
     from DataBase.conversation_store import ConversationStore  # type: ignore
-from .config import Settings, get_openai_api_key
-from .dialogue import DialogueManager, DialogState
-from .utils.container import get_container
+from ..config import Settings, get_openai_api_key
+from ..dialogue import DialogueManager, DialogState
+from ..utils.container import get_container
 
 
 def summarize_history(prev_summary: str, msgs: List[Dict[str, str]]) -> str:
