+++ OcchioOnniveggente/src/main.py
@@ -40,7 +40,7 @@ from src.hotword import is_wake
 from src.chat import ChatState
 from src.conversation import ConversationManager, DialogState
 from src.logging_utils import setup_logging
-from src.language_session import update_language
+from src.conversation import update_language
 from src.cli import _ensure_utf8_stdout, say, oracle_greeting, default_response
 from src.audio_device import pick_device, debug_print_devices
 from src.profile_utils import get_active_profile, make_domain_settings
